Changes were made

+++ c-rush-02/srcs/ft_lib/ft_itoa.c
@@ -15,7 +15,7 @@
 
 char	*ft_itoa(int nb)
 {
-	long	n;
+	int		n;
 	int		i;
 	char	*str;
